Release 1.1.14 (tracks 9abe66b3f4b9ec92db23af9cd0bc588371f9e4a7)

+++ dagster-cloud/dagster_cloud/storage/event_logs/queries.py
@@ -265,7 +265,7 @@ GET_MATERIALIZATION_COUNT_BY_PARTITION = """
     """
 
 GET_EVENT_TAGS_FOR_ASSET = """
-query getAssetEventTags($assetKey: String!, $filterTags: [AssetFilterTagInput!], $filterEventId: Int) {
+query getAssetEventTags($assetKey: String!, $filterTags: [AssetFilterTagInput!], $filterEventId: BigInt) {
     eventLogs {
         getAssetEventTags(assetKey: $assetKey, filterTags: $filterTags, filterEventId: $filterEventId) {
             ... on AssetEventTags {
